add update history

+++ initdb/create.sql
@@ -9,7 +9,7 @@ CREATE TABLE transferHistory (
   `to` varchar(20) NOT NULL,
   `toScan` varchar(200) NOT NULL,
   `token` varchar(20) NOT NULL,
-  `amount` int(18) NOT NULL,
+  `amount` float(18) NOT NULL,
   `mode` varchar(10) NOT NULL,
   `status` varchar(10) NOT NULL,
   `cratetime` timestamp NULL DEFAULT CURRENT_TIMESTAMP,
